implement an AsyncExecutor, get completion working
