+++ lua/cmp_obsidian.lua
@@ -19,58 +19,64 @@ source.complete = function(self, request, callback)
   local can_complete, search, insert_start, insert_end = completion.can_complete(request)
 
   if can_complete and search ~= nil and #search >= opts.completion.min_chars then
-    local items = {}
-    for note in client:search(search, "--ignore-case") do
-      local aliases = util.unique { tostring(note.id), note:display_name(), unpack(note.aliases) }
-      for _, alias in pairs(aliases) do
-        local options = {}
+    local function search_callback(results)
+      local items = {}
+      for _, result in ipairs(results) do
+        local note = result[1]
+        local aliases = util.unique { tostring(note.id), note:display_name(), unpack(note.aliases) }
+        for _, alias in pairs(aliases) do
+          local options = {}
 
-        local alias_case_matched = util.match_case(search, alias)
-        if
-          alias_case_matched ~= nil
-          and alias_case_matched ~= alias
-          and not util.contains(note.aliases, alias_case_matched)
-        then
-          table.insert(options, alias_case_matched)
-        end
-
-        table.insert(options, alias)
-
-        for _, option in pairs(options) do
-          local label = "[[" .. tostring(note.id)
-          if option ~= tostring(note.id) then
-            label = label .. "|" .. option .. "]]"
-          else
-            label = label .. "]]"
+          local alias_case_matched = util.match_case(search, alias)
+          if
+            alias_case_matched ~= nil
+            and alias_case_matched ~= alias
+            and not util.contains(note.aliases, alias_case_matched)
+          then
+            table.insert(options, alias_case_matched)
           end
 
-          table.insert(items, {
-            sortText = "[[" .. option,
-            label = label,
-            kind = 18,
-            textEdit = {
-              newText = label,
-              range = {
-                start = {
-                  line = request.context.cursor.row - 1,
-                  character = insert_start,
-                },
-                ["end"] = {
-                  line = request.context.cursor.row - 1,
-                  character = insert_end,
+          table.insert(options, alias)
+
+          for _, option in pairs(options) do
+            local label = "[[" .. tostring(note.id)
+            if option ~= tostring(note.id) then
+              label = label .. "|" .. option .. "]]"
+            else
+              label = label .. "]]"
+            end
+
+            table.insert(items, {
+              sortText = "[[" .. option,
+              label = label,
+              kind = 18,
+              textEdit = {
+                newText = label,
+                range = {
+                  start = {
+                    line = request.context.cursor.row - 1,
+                    character = insert_start,
+                  },
+                  ["end"] = {
+                    line = request.context.cursor.row - 1,
+                    character = insert_end,
+                  },
                 },
               },
-            },
-          })
+            })
+          end
         end
       end
+
+      callback {
+        items = items,
+        isIncomplete = false,
+      }
     end
-    return callback {
-      items = items,
-      isIncomplete = false,
-    }
+
+    client:search_async(search, "--ignore-case", search_callback)
   else
-    return callback { isIncomplete = true }
+    callback { isIncomplete = true }
   end
 end
